+++ packages/create-sitecore-jss/src/templates/react/scripts/generate-config.js
@@ -9,13 +9,11 @@ const { jssConfigFactory } = require('./config');
 /**
  * REACT_APP is the default prefix for environment variables that will be made available to the app.
  */
-// REACT_APP_JSS_APP_NAME env variable has been deprecated since v.21.6, REACT_APP_SITE_NAME should be used instead
 const defaultConfig = {
   sitecoreApiKey: process.env[`${constantCase('reactAppSitecoreApiKey')}`],
   sitecoreApiHost: process.env[`${constantCase('reactAppSitecoreApiHost')}`],
   sitecoreSiteName:
-    process.env[`${constantCase('reactAppSitecoreSiteName')}`] ||
-    process.env[`${constantCase('reactAppJssAppName')}`],
+    process.env[`${constantCase('reactAppSitecoreSiteName')}`],
   graphQLEndpointPath: process.env[`${constantCase('reactAppGraphQLEndpointPath')}`],
   defaultLanguage: process.env[`${constantCase('reactAppDefaultLanguage')}`],
   graphQLEndpoint: process.env[`${constantCase('reactAppGraphQLEndpoint')}`],