+++ packages/sitecore-jss-nextjs/src/services/mutisite-graphql-sitemap-service.test.ts
@@ -22,6 +22,10 @@ describe('MultisiteGraphQLSitemapService', () => {
   const endpoint = 'http://site';
   const apiKey = 'some-api-key';
   const sites = ['site-name'];
+  const clientFactory = GraphQLRequestClient.createClientFactory({
+    endpoint,
+    apiKey,
+  });
 
   afterEach(() => {
     nock.cleanAll();
@@ -56,7 +60,7 @@ describe('MultisiteGraphQLSitemapService', () => {
     it('should work when 1 language is requested', async () => {
       mockPathsRequest();
 
-      const service = new MultisiteGraphQLSitemapService({ endpoint, apiKey, sites });
+      const service = new MultisiteGraphQLSitemapService({ clientFactory, sites });
       const sitemap = await service.fetchSSGSitemap(['ua']);
       expect(sitemap).to.deep.equal(sitemapServiceMultisiteResult);
 
@@ -123,8 +127,7 @@ describe('MultisiteGraphQLSitemapService', () => {
         });
 
       const service = new MultisiteGraphQLSitemapService({
-        endpoint,
-        apiKey,
+        clientFactory,
         sites,
         includedPaths,
         excludedPaths,
@@ -145,7 +148,7 @@ describe('MultisiteGraphQLSitemapService', () => {
       it('should work when 1 language is requested', async () => {
         mockPathsRequest();
 
-        const service = new MultisiteGraphQLSitemapService({ endpoint, apiKey, sites });
+        const service = new MultisiteGraphQLSitemapService({ clientFactory, sites });
         const sitemap = await service.fetchSSGSitemap(['ua']);
         expect(sitemap).to.deep.equal(sitemapServiceMultisiteResult);
 
@@ -212,8 +215,7 @@ describe('MultisiteGraphQLSitemapService', () => {
           });
 
         const service = new MultisiteGraphQLSitemapService({
-          endpoint,
-          apiKey,
+          clientFactory,
           sites,
           includedPaths,
           excludedPaths,
@@ -297,8 +299,7 @@ describe('MultisiteGraphQLSitemapService', () => {
           });
 
         const service = new MultisiteGraphQLSitemapService({
-          endpoint,
-          apiKey,
+          clientFactory,
           sites: multipleSites,
         });
         const sitemap = await service.fetchSSGSitemap([lang]);
@@ -387,8 +388,7 @@ describe('MultisiteGraphQLSitemapService', () => {
           });
 
         const service = new MultisiteGraphQLSitemapService({
-          endpoint,
-          apiKey,
+          clientFactory,
           sites: multipleSites,
           includePersonalizedRoutes: true,
         });
@@ -529,7 +529,7 @@ describe('MultisiteGraphQLSitemapService', () => {
             },
           });
 
-        const service = new MultisiteGraphQLSitemapService({ endpoint, apiKey, sites });
+        const service = new MultisiteGraphQLSitemapService({ clientFactory, sites });
         const sitemap = await service.fetchSSGSitemap([lang1, lang2]);
 
         expect(sitemap).to.deep.equal([
@@ -618,7 +618,7 @@ describe('MultisiteGraphQLSitemapService', () => {
             },
           });
 
-        const service = new MultisiteGraphQLSitemapService({ endpoint, apiKey, sites });
+        const service = new MultisiteGraphQLSitemapService({ clientFactory, sites });
         const sitemap = await service.fetchSSGSitemap([lang]);
 
         expect(sitemap).to.deep.equal([
@@ -640,14 +640,14 @@ describe('MultisiteGraphQLSitemapService', () => {
       });
 
       it('should throw error if valid language is not provided', async () => {
-        const service = new MultisiteGraphQLSitemapService({ endpoint, apiKey, sites });
+        const service = new MultisiteGraphQLSitemapService({ clientFactory, sites });
         await service.fetchSSGSitemap([]).catch((error: RangeError) => {
           expect(error.message).to.equal(languageError);
         });
       });
 
       it('should throw error if query returns nothing for a provided site name', async () => {
-        const service = new MultisiteGraphQLSitemapService({ endpoint, apiKey, sites });
+        const service = new MultisiteGraphQLSitemapService({ clientFactory, sites });
         nock(endpoint)
           .post('/', (body) => {
             return body.variables.siteName === sites[0];
@@ -667,7 +667,7 @@ describe('MultisiteGraphQLSitemapService', () => {
       it('should throw error if empty language is provided', async () => {
         mockPathsRequest();
 
-        const service = new MultisiteGraphQLSitemapService({ endpoint, apiKey, sites });
+        const service = new MultisiteGraphQLSitemapService({ clientFactory, sites });
         await service.fetchExportSitemap('').catch((error: RangeError) => {
           expect(error.message).to.equal('The language must be a non-empty string');
         });
@@ -683,8 +683,7 @@ describe('MultisiteGraphQLSitemapService', () => {
           .reply(200, sitemapDefaultQueryResult);
 
         const service = new MultisiteGraphQLSitemapService({
-          endpoint,
-          apiKey,
+          clientFactory,
           sites,
           pageSize: customPageSize,
         });
@@ -704,8 +703,7 @@ describe('MultisiteGraphQLSitemapService', () => {
           .reply(200, sitemapDefaultQueryResult);
 
         const service = new MultisiteGraphQLSitemapService({
-          endpoint,
-          apiKey,
+          clientFactory,
           sites,
           pageSize: undefined,
         });
@@ -718,7 +716,7 @@ describe('MultisiteGraphQLSitemapService', () => {
       it('should work if sitemap has 0 pages', async () => {
         mockPathsRequest([]);
 
-        const service = new MultisiteGraphQLSitemapService({ endpoint, apiKey, sites });
+        const service = new MultisiteGraphQLSitemapService({ clientFactory, sites });
         const sitemap = await service.fetchSSGSitemap(['ua']);
         expect(sitemap).to.deep.equal([]);
         return expect(nock.isDone()).to.be.true;
@@ -729,7 +727,7 @@ describe('MultisiteGraphQLSitemapService', () => {
           .post('/', /DefaultSitemapQuery/gi)
           .reply(500, 'Error 😥');
 
-        const service = new MultisiteGraphQLSitemapService({ endpoint, apiKey, sites });
+        const service = new MultisiteGraphQLSitemapService({ clientFactory, sites });
         await service.fetchSSGSitemap(['ua']).catch((error: RangeError) => {
           expect(error.message).to.contain('SitemapQuery');
           expect(error.message).to.contain('Error 😥');
@@ -765,14 +763,14 @@ describe('MultisiteGraphQLSitemapService', () => {
   describe('Fetch sitemap in export mode', () => {
     it('should fetch multisite sitemap', async () => {
       mockPathsRequest();
-      const service = new MultisiteGraphQLSitemapService({ endpoint, apiKey, sites });
+      const service = new MultisiteGraphQLSitemapService({ clientFactory, sites });
       const sitemap = await service.fetchExportSitemap('ua');
       expect(sitemap).to.deep.equal(expectedMultisiteExportSitemap);
       return expect(nock.isDone()).to.be.true;
     });
     it('should work if endpoint returns 0 pages', async () => {
       mockPathsRequest([]);
-      const service = new MultisiteGraphQLSitemapService({ endpoint, apiKey, sites });
+      const service = new MultisiteGraphQLSitemapService({ clientFactory, sites });
       const sitemap = await service.fetchExportSitemap('ua');
       expect(sitemap).to.deep.equal([]);
       return expect(nock.isDone()).to.be.true;
@@ -781,7 +779,7 @@ describe('MultisiteGraphQLSitemapService', () => {
       nock(endpoint)
         .post('/', /DefaultSitemapQuery/gi)
         .reply(500, 'Error 😥');
-      const service = new MultisiteGraphQLSitemapService({ endpoint, apiKey, sites });
+      const service = new MultisiteGraphQLSitemapService({ clientFactory, sites });
       await service.fetchExportSitemap('ua').catch((error: RangeError) => {
         expect(error.message).to.contain('SitemapQuery');
         expect(error.message).to.contain('Error 😥');
@@ -790,7 +788,7 @@ describe('MultisiteGraphQLSitemapService', () => {
     });
     it('should throw error if language is not provided', async () => {
       mockPathsRequest();
-      const service = new MultisiteGraphQLSitemapService({ endpoint, apiKey, sites });
+      const service = new MultisiteGraphQLSitemapService({ clientFactory, sites });
       await service.fetchExportSitemap('').catch((error: RangeError) => {
         expect(error.message).to.equal('The language must be a non-empty string');
       });
@@ -798,7 +796,7 @@ describe('MultisiteGraphQLSitemapService', () => {
     });
   });
   it('should throw error if query returns nothing for a provided site name', async () => {
-    const service = new MultisiteGraphQLSitemapService({ endpoint, apiKey, sites });
+    const service = new MultisiteGraphQLSitemapService({ clientFactory, sites });
     nock(endpoint)
       .post('/', (body) => {
         return body.variables.siteName === sites[0];
@@ -814,10 +812,10 @@ describe('MultisiteGraphQLSitemapService', () => {
       expect(error.message).to.equal(getSiteEmptyError(sites[0]));
     });
   });
+
   it('should provide a default GraphQL client', () => {
     const service = new TestService({
-      endpoint,
-      apiKey,
+      clientFactory: clientFactory,
       sites,
     });
     const graphQLClient = service.client as GraphQLClient;