+++ packages/sitecore-jss/src/site/graphql-robots-service.ts
@@ -1,4 +1,4 @@
-import { GraphQLClient, GraphQLRequestClient } from '../graphql';
+import { GraphQLClient } from '../graphql';
 import { siteNameError } from '../constants';
 import debug from '../debug';
 import { GraphQLRequestClientFactory } from '../graphql-request-client';
@@ -16,16 +16,6 @@ const defaultQuery = /* GraphQL */ `
 
 export type GraphQLRobotsServiceConfig = {
   /**
-   * Your Graphql endpoint
-   * @deprecated use @param clientFactory property instead
-   */
-  endpoint?: string;
-  /**
-   * The API key to use for authentication
-   * @deprecated use @param clientFactory property instead
-   */
-  apiKey?: string;
-  /**
    * The JSS application name
    */
   siteName: string;
@@ -33,7 +23,7 @@ export type GraphQLRobotsServiceConfig = {
    * A GraphQL Request Client Factory is a function that accepts configuration and returns an instance of a GraphQLRequestClient.
    * This factory function is used to create and configure GraphQL clients for making GraphQL API requests.
    */
-  clientFactory?: GraphQLRequestClientFactory;
+  clientFactory: GraphQLRequestClientFactory;
 };
 
 /**
@@ -90,18 +80,11 @@ export class GraphQLRobotsService {
    * @returns {GraphQLClient} implementation
    */
   protected getGraphQLClient(): GraphQLClient {
-    if (!this.options.endpoint) {
-      if (!this.options.clientFactory) {
-        throw new Error('You should provide either an endpoint and apiKey, or a clientFactory.');
-      }
-
-      return this.options.clientFactory({
-        debugger: debug.robots,
-      });
+    if (!this.options.clientFactory) {
+      throw new Error('clientFactory needs to be provided when initializing GraphQL client.');
     }
 
-    return new GraphQLRequestClient(this.options.endpoint, {
-      apiKey: this.options.apiKey,
+    return this.options.clientFactory({
       debugger: debug.robots,
     });
   }